Add hover text change event over languages
Feature:
When a user's mouse is hovered over any one of the four languages
the text changes.

What I did:
- add nowrap feature in css
- remove duplicate button selector
- add inline onmouseover and onmouseout events to language buttons

What I learned:
- How to change text on hover
- How to have the korean text on one line

I learned how to finally change text on hover. When I initially created
this website, during the beginning of the month, I wanted to have text
change into the language listed on the buttons. I was not able to
accomplish that, so I gave up and settled for a raised button on hover
event.

I still ran into a problem. When I hovered over Korean it would change
to 한국어, but each block would be it's own seperate line. I then learned
about the nowrap feature.